LayoutTests/webaudio/silent-audio-interrupted-in-background-expected.txt

Tests silent WebAudio rendering in hidden pages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.state became 'running'
* Setting page visibility to hidden
PASS document.hidden is true
PASS context.state is 'running'
* Setting page visibility to visible
PASS document.hidden is false
PASS context.state is 'running'
PASS successfullyParsed is true

TEST COMPLETE
